extract provision logic form zos node into its own package
this leaves the provision package as a generic package that can be used
by any other party to implement new primitives type and provisioning
logic
