+++ pkg/provision/primitives/counters.go
@@ -1,0 +1,232 @@
+package primitives
+
+import (
+	"encoding/json"
+	"fmt"
+	"sync/atomic"
+
+	"github.com/threefoldtech/tfexplorer/models/generated/directory"
+	"github.com/threefoldtech/zos/pkg/provision"
+)
+
+// Counter interface
+type Counter interface {
+	// Increment counter atomically by v
+	Increment(v uint64) uint64
+	// Decrement counter atomically by v
+	Decrement(v uint64) uint64
+	// Current returns the current value
+	Current() uint64
+}
+
+type counterNop struct{}
+
+func (c *counterNop) Increment(v uint64) uint64 {
+	return 0
+}
+
+func (c *counterNop) Decrement(v uint64) uint64 {
+	return 0
+}
+
+func (c *counterNop) Current() uint64 {
+	return 0
+}
+
+// counterImpl value for safe increment/decrement
+type counterImpl uint64
+
+// Increment counter atomically by one
+func (c *counterImpl) Increment(v uint64) uint64 {
+	return atomic.AddUint64((*uint64)(c), v)
+}
+
+// Decrement counter atomically by one
+func (c *counterImpl) Decrement(v uint64) uint64 {
+	return atomic.AddUint64((*uint64)(c), -v)
+}
+
+// Current returns the current value
+func (c *counterImpl) Current() uint64 {
+	return atomic.LoadUint64((*uint64)(c))
+}
+
+// Counters tracks the amount of primitives workload deployed and
+// the amount of resource unit used
+type Counters struct {
+	containers counterImpl
+	volumes    counterImpl
+	networks   counterImpl
+	zdbs       counterImpl
+	vms        counterImpl
+	debugs     counterImpl
+
+	SRU counterImpl // SSD storage in bytes
+	HRU counterImpl // HDD storage in bytes
+	MRU counterImpl // Memory storage in bytes
+	CRU counterImpl // CPU count absolute
+}
+
+func (c *Counters) CurrentWorkloads() directory.WorkloadAmount {
+	return directory.WorkloadAmount{
+		Network:      uint16(c.containers.Current()),
+		Volume:       uint16(c.volumes.Current()),
+		ZDBNamespace: uint16(c.networks.Current()),
+		Container:    uint16(c.zdbs.Current()),
+		K8sVM:        uint16(c.vms.Current()),
+	}
+}
+
+func (c *Counters) CurrentUnits() directory.ResourceAmount {
+	return directory.ResourceAmount{
+		Cru: c.CRU.Current(),
+		Mru: float64(c.CRU.Current()),
+		Hru: float64(c.CRU.Current()),
+		Sru: float64(c.CRU.Current()),
+	}
+}
+
+const (
+	mib = uint64(1024 * 1024)
+	gib = uint64(mib * 1024)
+)
+
+func (c *Counters) Increment(r *provision.Reservation) error {
+
+	var (
+		u   resourceUnits
+		err error
+	)
+
+	switch r.Type {
+	case VolumeReservation:
+		u, err = processVolume(r)
+	case ContainerReservation:
+		u, err = processContainer(r)
+	case ZDBReservation:
+		u, err = processZdb(r)
+	case KubernetesReservation:
+		c.vms.Increment(1)
+		u, err = processKubernetes(r)
+	}
+	if err != nil {
+		return err
+	}
+
+	c.CRU.Increment(u.CRU)
+	c.MRU.Increment(u.MRU)
+	c.SRU.Increment(u.SRU)
+	c.HRU.Increment(u.HRU)
+
+	return nil
+}
+
+func (c *Counters) Decrement(r *provision.Reservation) error {
+
+	var (
+		u   resourceUnits
+		err error
+	)
+
+	switch r.Type {
+	case VolumeReservation:
+		c.volumes.Decrement(1)
+		u, err = processVolume(r)
+	case ContainerReservation:
+		c.volumes.Decrement(1)
+		u, err = processContainer(r)
+	case ZDBReservation:
+		c.zdbs.Decrement(1)
+		u, err = processZdb(r)
+	case KubernetesReservation:
+		c.vms.Decrement(1)
+		u, err = processKubernetes(r)
+	}
+	if err != nil {
+		return err
+	}
+
+	c.CRU.Decrement(u.CRU)
+	c.MRU.Decrement(u.MRU)
+	c.SRU.Decrement(u.SRU)
+	c.HRU.Decrement(u.HRU)
+
+	return nil
+}
+
+type resourceUnits struct {
+	SRU uint64 `json:"sru,omitempty"`
+	HRU uint64 `json:"hru,omitempty"`
+	MRU uint64 `json:"mru,omitempty"`
+	CRU uint64 `json:"cru,omitempty"`
+}
+
+func processVolume(r *provision.Reservation) (u resourceUnits, err error) {
+	var volume Volume
+	if err = json.Unmarshal(r.Data, &volume); err != nil {
+		return u, err
+	}
+
+	// volume.size and SRU is in GiB
+	switch volume.Type {
+	case SSDDiskType:
+		u.SRU = volume.Size * gib
+	case HDDDiskType:
+		u.HRU = volume.Size * gib
+	}
+
+	return u, nil
+}
+
+func processContainer(r *provision.Reservation) (u resourceUnits, err error) {
+	var cont Container
+	if err = json.Unmarshal(r.Data, &cont); err != nil {
+		return u, err
+	}
+	u.CRU = uint64(cont.Capacity.CPU)
+	// memory is in MiB
+	u.MRU = cont.Capacity.Memory * mib
+	u.SRU = 256 * mib // 250MiB are allocated on SSD for the root filesystem used by the flist
+
+	return u, nil
+}
+
+func processZdb(r *provision.Reservation) (u resourceUnits, err error) {
+	if r.Type != ZDBReservation {
+		return u, fmt.Errorf("wrong type or reservation %s, excepted %s", r.Type, ZDBReservation)
+	}
+	var zdbVolume ZDB
+	if err := json.Unmarshal(r.Data, &zdbVolume); err != nil {
+		return u, err
+	}
+
+	switch zdbVolume.DiskType {
+	case "SSD":
+		u.SRU = zdbVolume.Size * gib
+	case "HDD":
+		u.HRU = zdbVolume.Size * gib
+	}
+
+	return u, nil
+}
+
+func processKubernetes(r *provision.Reservation) (u resourceUnits, err error) {
+	var k8s Kubernetes
+	if err = json.Unmarshal(r.Data, &k8s); err != nil {
+		return u, err
+	}
+
+	// size are defined at https://github.com/threefoldtech/zos/blob/master/pkg/provision/kubernetes.go#L311
+	switch k8s.Size {
+	case 1:
+		u.CRU = 1
+		u.MRU = 2 * gib
+		u.SRU = 50 * gib
+	case 2:
+		u.CRU = 2
+		u.MRU = 4 * gib
+		u.SRU = 100 * gib
+	}
+
+	return u, nil
+}
