+++ pkg/provision/primitives/network_test.go
@@ -1,4 +1,4 @@
-package provision
+package primitives
 
 import (
 	"context"
