+++ pkg/provision/reservation.go
@@ -10,37 +10,10 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/threefoldtech/tfexplorer/models/generated/workloads"
-	"github.com/threefoldtech/tfexplorer/schema"
-	"github.com/threefoldtech/zos/pkg"
-	"github.com/threefoldtech/zos/pkg/container/logger"
-	"github.com/threefoldtech/zos/pkg/container/stats"
-	"github.com/threefoldtech/zos/pkg/versioned"
 )
 
 // ReservationType type
 type ReservationType string
-
-const (
-	// ContainerReservation type
-	ContainerReservation ReservationType = "container"
-	// VolumeReservation type
-	VolumeReservation ReservationType = "volume"
-	// NetworkReservation type
-	NetworkReservation ReservationType = "network"
-	// ZDBReservation type
-	ZDBReservation ReservationType = "zdb"
-	// DebugReservation type
-	DebugReservation ReservationType = "debug"
-	// KubernetesReservation type
-	KubernetesReservation ReservationType = "kubernetes"
-)
-
-var (
-	// reservationSchemaV1 reservation schema version 1
-	reservationSchemaV1 = versioned.MustParse("1.0.0")
-	// reservationSchemaLastVersion link to latest version
-	reservationSchemaLastVersion = reservationSchemaV1
-)
 
 // Reservation struct
 type Reservation struct {
@@ -148,179 +121,6 @@ func (r *Reservation) validate() error {
 	}
 
 	return nil
-}
-
-func networkReservation(i interface{}) workloads.Network {
-	n := i.(pkg.Network)
-	network := workloads.Network{
-		Name:             n.Name,
-		Iprange:          n.IPRange.ToSchema(),
-		WorkloadId:       1,
-		NetworkResources: make([]workloads.NetworkNetResource, len(n.NetResources)),
-	}
-
-	for i, nr := range n.NetResources {
-		network.NetworkResources[i] = workloads.NetworkNetResource{
-			NodeId:                       nr.NodeID,
-			Iprange:                      nr.Subnet.ToSchema(),
-			WireguardPrivateKeyEncrypted: nr.WGPrivateKey,
-			WireguardPublicKey:           nr.WGPublicKey,
-			WireguardListenPort:          int64(nr.WGListenPort),
-			Peers:                        make([]workloads.WireguardPeer, len(nr.Peers)),
-		}
-
-		for y, peer := range nr.Peers {
-			network.NetworkResources[i].Peers[y] = workloads.WireguardPeer{
-				Iprange:        peer.Subnet.ToSchema(),
-				Endpoint:       peer.Endpoint,
-				PublicKey:      peer.WGPublicKey,
-				AllowedIprange: make([]schema.IPRange, len(peer.AllowedIPs)),
-			}
-
-			for z, ip := range peer.AllowedIPs {
-				network.NetworkResources[i].Peers[y].AllowedIprange[z] = ip.ToSchema()
-			}
-		}
-	}
-	return network
-}
-
-func containerReservation(i interface{}, nodeID string) workloads.Container {
-
-	c := i.(Container)
-	container := workloads.Container{
-		NodeId:            nodeID,
-		WorkloadId:        1,
-		Flist:             c.FList,
-		HubUrl:            c.FlistStorage,
-		Environment:       c.Env,
-		SecretEnvironment: c.SecretEnv,
-		Entrypoint:        c.Entrypoint,
-		Interactive:       c.Interactive,
-		Volumes:           make([]workloads.ContainerMount, len(c.Mounts)),
-		StatsAggregator:   make([]workloads.StatsAggregator, len(c.StatsAggregator)),
-		Logs:              make([]workloads.Logs, len(c.Logs)),
-		NetworkConnection: []workloads.NetworkConnection{
-			{
-				NetworkId: string(c.Network.NetworkID),
-				Ipaddress: c.Network.IPs[0],
-				PublicIp6: c.Network.PublicIP6,
-			},
-		},
-		Capacity: workloads.ContainerCapacity{
-			Cpu:    int64(c.Capacity.CPU),
-			Memory: int64(c.Capacity.Memory),
-		},
-	}
-
-	for i, v := range c.Mounts {
-		container.Volumes[i] = workloads.ContainerMount{
-			VolumeId:   v.VolumeID,
-			Mountpoint: v.Mountpoint,
-		}
-	}
-
-	for i, l := range c.Logs {
-		if l.Type != logger.RedisType {
-			container.Logs[i] = workloads.Logs{
-				Type: "unknown",
-				Data: workloads.LogsRedis{},
-			}
-
-			continue
-		}
-
-		container.Logs[i] = workloads.Logs{
-			Type: l.Type,
-			Data: workloads.LogsRedis{
-				Stdout: l.Data.Stdout,
-				Stderr: l.Data.Stderr,
-			},
-		}
-	}
-
-	for i, s := range c.StatsAggregator {
-		if s.Type != stats.RedisType {
-			container.StatsAggregator[i] = workloads.StatsAggregator{
-				Type: "unknown",
-				Data: workloads.StatsRedis{},
-			}
-
-			continue
-		}
-
-		container.StatsAggregator[i] = workloads.StatsAggregator{
-			Type: s.Type,
-			Data: workloads.StatsRedis{
-				Endpoint: s.Data.Endpoint,
-			},
-		}
-	}
-
-	return container
-}
-
-func volumeReservation(i interface{}, nodeID string) workloads.Volume {
-	v := i.(Volume)
-
-	volume := workloads.Volume{
-		NodeId:     nodeID,
-		WorkloadId: 1,
-		Size:       int64(v.Size),
-	}
-
-	if v.Type == HDDDiskType {
-		volume.Type = workloads.VolumeTypeHDD
-	} else if v.Type == SSDDiskType {
-		volume.Type = workloads.VolumeTypeSSD
-	}
-
-	return volume
-}
-
-func zdbReservation(i interface{}, nodeID string) workloads.ZDB {
-	z := i.(ZDB)
-
-	zdb := workloads.ZDB{
-		WorkloadId: 1,
-		NodeId:     nodeID,
-		// ReservationID:
-		Size:     int64(z.Size),
-		Password: z.Password,
-		Public:   z.Public,
-		// StatsAggregator:
-		// FarmerTid:
-	}
-	if z.DiskType == pkg.SSDDevice {
-		zdb.DiskType = workloads.DiskTypeHDD
-	} else if z.DiskType == pkg.HDDDevice {
-		zdb.DiskType = workloads.DiskTypeSSD
-	}
-
-	if z.Mode == pkg.ZDBModeUser {
-		zdb.Mode = workloads.ZDBModeUser
-	} else if z.Mode == pkg.ZDBModeSeq {
-		zdb.Mode = workloads.ZDBModeSeq
-	}
-
-	return zdb
-}
-
-func k8sReservation(i interface{}, nodeID string) workloads.K8S {
-	k := i.(Kubernetes)
-
-	k8s := workloads.K8S{
-		WorkloadId:    1,
-		NodeId:        nodeID,
-		Size:          int64(k.Size),
-		NetworkId:     string(k.NetworkID),
-		Ipaddress:     k.IP,
-		ClusterSecret: k.ClusterSecret,
-		MasterIps:     k.MasterIPs,
-		SshKeys:       k.SSHKeys,
-	}
-
-	return k8s
 }
 
 // ResultState type
